chore: Refactor SettingsMenu into separate component. (#1392)
Refactor SettingsMenu into separate component, include in component context
