+++ ui/src/components/NavBar/index.ts
@@ -1,0 +1,1 @@
+export { withNavBar } from "./NavBar";
