+++ ui/src/components/NavBar/nav-bar.module.scss
@@ -101,7 +101,3 @@ $bg-color-transition: background-color 0.2s ease;
   width: 150px;
   margin-bottom: -8px;
 }
-
-.settings-icon {
-  margin-right: 8px;
-}
