+++ ui/src/components/SettingsMenu/index.ts
@@ -1,0 +1,3 @@
+export { SettingsMenu } from "./SettingsMenu";
+
+export type { SettingsMenuProps } from "./SettingsMenu";
