+++ ui/src/components/SettingsMenu/settings-menu.module.scss
@@ -1,0 +1,3 @@
+.settings-icon {
+  margin-right: 8px;
+}
